feat(functools): add submodule `optree.functools`
