+++ docs/source/functools.rst
@@ -1,0 +1,12 @@
+PyTree integration with :mod:`functools.partial`
+================================================
+
+.. currentmodule:: optree.functools
+
+.. autosummary::
+
+    partial
+    reduce
+
+.. autoclass:: partial
+.. autofunction:: reduce
