+++ docs/source/registry.rst
@@ -8,7 +8,6 @@ PyTree Node Registry
     register_pytree_node
     register_pytree_node_class
     unregister_pytree_node
-    Partial
     register_keypaths
     AttributeKeyPathEntry
     GetitemKeyPathEntry
@@ -16,7 +15,6 @@ PyTree Node Registry
 .. autofunction:: register_pytree_node
 .. autofunction:: register_pytree_node_class
 .. autofunction:: unregister_pytree_node
-.. autofunction:: Partial
 .. autofunction:: register_keypaths
 .. autofunction:: AttributeKeyPathEntry
 .. autofunction:: GetitemKeyPathEntry
