+++ docs/source/spelling_wordlist.txt
@@ -75,6 +75,7 @@ jax
 numpy
 torch
 dtype
+cuda
 getattr
 setattr
 delattr
